Revert "Add an option to support Feature Management V2 schema (#316)" (#328)
This reverts commit 7233900221a3884bae70acabd048c61033708e63.
