--- src/Microsoft.Extensions.Configuration.AzureAppConfiguration/FeatureManagement/DynamicFeature.cs
@@ -1,20 +1,0 @@
-﻿// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT license.
-//
-using System.Collections.Generic;
-using System.Text.Json.Serialization;
-
-namespace Microsoft.Extensions.Configuration.AzureAppConfiguration.FeatureManagement
-{
-    internal class DynamicFeature
-    {
-        [JsonPropertyName("id")]
-        public string Id { get; set; }
-
-        [JsonPropertyName("client_assigner")]
-        public string ClientAssigner { get; set; }
-
-        [JsonPropertyName("variants")]
-        public List<FeatureVariant> Variants { get; set; } = new List<FeatureVariant>();
-    }
-}
