+++ src/Microsoft.Extensions.Configuration.AzureAppConfiguration/RequestTracingOptions.cs
@@ -1,8 +1,6 @@
 ﻿// Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 //
-using Microsoft.Extensions.Configuration.AzureAppConfiguration.FeatureManagement;
-
 namespace Microsoft.Extensions.Configuration.AzureAppConfiguration
 {
     internal class RequestTracingOptions
@@ -26,10 +24,5 @@ namespace Microsoft.Extensions.Configuration.AzureAppConfiguration
         /// Flag to indicate whether the request is from a development environment.
         /// </summary>
         public bool IsDevEnvironment { get; set; } = false;
-
-        /// <summary>
-        /// This value indicates the feature management schema version being used. 
-        /// </summary>
-        public string FeatureManagementSchemaVersion { get; set; }
     }
 }
